Working ascent trajectory!

+++ RocketOptimize/Simulation/Integrators/Integrator.cs
@@ -7,6 +7,6 @@ namespace RocketOptimize.Simulation.Integrators
 
     public interface IIntegrator
     {
-        void Integrate(float timeStep, ref State currentState, out State newState, AccelerationCalculator calculateAcceleration);
+        void Integrate(double timeStep, ref State currentState, out State newState, AccelerationCalculator calculateAcceleration);
     }
 }
