+++ RocketOptimize/Simulation/State.cs
@@ -4,7 +4,7 @@ namespace RocketOptimize.Simulation
 {
     public struct State
     {
-        public float Time;
+        public double Time;
         public Vector3d Position;
         public Vector3d Velocity;
         public Vector3d Acceleration;
